Move price symbol policy to provider domain
